[spaces.dg.pdelab] moved DiscontinuousLangrange -> DG, refs #14
